+++ Code/Mantid/Framework/API/inc/MantidAPI/WorkspaceFactory.h
@@ -62,9 +62,9 @@ class EXPORT_OPT_MANTID_API WorkspaceFactoryImpl : public Kernel::DynamicFactory
 {
 public:
   MatrixWorkspace_sptr create(const MatrixWorkspace_const_sptr& parent,
-                        int64_t NVectors = -1, int64_t XLength = -1, int64_t YLength = -1) const;
-  MatrixWorkspace_sptr create(const std::string& className, const int64_t& NVectors,
-                                   const int64_t& XLength, const int64_t& YLength) const;
+                                   size_t NVectors = size_t(-1), size_t XLength = size_t(-1), size_t YLength = size_t(-1)) const;
+  MatrixWorkspace_sptr create(const std::string& className, const size_t& NVectors,
+                                   const size_t& XLength, const size_t& YLength) const;
   
   IMDWorkspace_sptr create(const std::string & className, const Geometry::MDGeometryDescription &) const;
  /// this create method is currently used to build MD workspaces, but may be used to build MD workspaces from matrix workspaces in a future;
